ford: cast the pinion angle to int before subtracting the CAN offset
((msg->data[2] & 0x7FU) << 8 | msg->data[3]) - 16000U is unsigned arithmetic
(MISRA C:2012 10.3/10.8): the left operand is unsigned int, subtracting
16000U wraps for every raw value below 16000 -- every negative angle, half
the signal's range -- and only reaches int through an implementation-defined
conversion. It happens to yield the right value on a two's-complement
target, which is why the existing tests already passed. Cast the extraction
to int before subtracting a signed 16000, matching the idiom already used
for similar 15-bit signals elsewhere (e.g. volvo.h's steering angle
extractions).

Add test_pinion_angle_decode_at_negative_angles, which pins the decoded
angle_meas value for raw pinion readings below the 16000 offset directly,
rather than relying only on a downstream bound outcome.

+++ opendbc_repo/opendbc/safety/modes/ford.h
@@ -210,7 +210,7 @@ static void ford_rx_hook(const CANPacket_t *msg) {
     if ((msg->addr == FORD_SteeringPinion_Data) && ford_lka_steering) {
       // Signal: StePinComp_An_Est : 22|15@0+ (0.1,-1600) degrees -> tenths of a degree,
       // matching FORD_LKA_DEG_TO_CAN in ford_tx_hook
-      const int pinion_angle = (((msg->data[2] & 0x7FU) << 8) | msg->data[3]) - 16000U;
+      const int pinion_angle = (int)(((msg->data[2] & 0x7FU) << 8) | msg->data[3]) - 16000;
       update_sample(&angle_meas, pinion_angle);
     }
